[EA3-118] fix: ReviewTags ReviewFinder 테스트 반대로 변경
- ReviewTags에 ReviewFinder 테스트를 작성한 문제
